+++ packages/camera/camera_web/example/integration_test/camera_service_test.dart
@@ -673,6 +673,60 @@ void main() {
         });
 
         testWidgets('returns null '
+            'when the facing mode setting is empty and '
+            'the first facingMode capability is not a JavaScript string', (
+          WidgetTester tester,
+        ) async {
+          mockVideoTrack.getSettings = () {
+            return web.MediaTrackSettings(facingMode: '');
+          }.toJS;
+          mockVideoTrack.getCapabilities = () {
+            return createJSInteropWrapper(FakeMediaTrackCapabilities())
+                as web.MediaTrackCapabilities;
+          }.toJS;
+
+          when(
+            jsUtil.hasProperty(videoTrack, 'getCapabilities'.toJS),
+          ).thenReturn(true);
+          when(
+            jsUtil.getProperty(any, 'facingMode'.toJS),
+          ).thenReturn(<JSAny>[true.toJS].toJS);
+
+          final String? facingMode = cameraService.getFacingModeForVideoTrack(
+            videoTrack,
+          );
+
+          expect(facingMode, isNull);
+        });
+
+        testWidgets('returns null '
+            'when the facing mode setting is empty and '
+            'the facingMode capability is not a JavaScript array', (
+          WidgetTester tester,
+        ) async {
+          mockVideoTrack.getSettings = () {
+            return web.MediaTrackSettings(facingMode: '');
+          }.toJS;
+          mockVideoTrack.getCapabilities = () {
+            return createJSInteropWrapper(FakeMediaTrackCapabilities())
+                as web.MediaTrackCapabilities;
+          }.toJS;
+
+          when(
+            jsUtil.hasProperty(videoTrack, 'getCapabilities'.toJS),
+          ).thenReturn(true);
+          when(
+            jsUtil.getProperty(any, 'facingMode'.toJS),
+          ).thenReturn(true.toJS);
+
+          final String? facingMode = cameraService.getFacingModeForVideoTrack(
+            videoTrack,
+          );
+
+          expect(facingMode, isNull);
+        });
+
+        testWidgets('returns null '
             'when the facing mode setting '
             'and capabilities are empty', (WidgetTester tester) async {
           mockVideoTrack.getSettings = () {